requirements.txt
# Microsoft Agent Framework (--pre flag required while in preview)
agent-framework-azure-ai

# Azure Authentication
azure-identity
monocle_apptrace